requirements.txt
apache-airflow-providers-http
apache-airflow-providers-smtp
apache-airflow-providers-postgres
